pyproject.toml
[project]
name = "ai-rpa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "markdown>=3.10.2",
    "openai>=2.31.0",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]
